DjangoWebProject1/requirements.txt
asgiref==3.7.2
astroid==3.1.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
dill==0.3.8
Django==5.0.2
django-bootstrap3==23.6
django-ckeditor==6.7.1
django-crispy-forms==1.11.2
django-environ==0.11.2
django-etc==1.3.0
django-hitcount==1.3.5
django-js-asset==2.2.0
django-precise-bbcode==1.2.16
django-resized==0.3.11
django-social-auth==0.7.28
django-taggit==5.0.1
django-tinymce==3.3.0
dj-database-url==2.1.0
eventemitter==0.2.0
fontawesomefree==6.5.1
gevent==24.2.1
google==3.0.0
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
install==1.3.5
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.6.1
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==24.0
Pillow==9.5.0
pip==22.3
platformdirs==4.2.0
psycopg2==2.9.9
pycparser==2.21
pycryptodomex==3.20.0
PyJWT==2.8.0
pylint==3.1.0
pyparsing==3.1.2
python3-openid==3.2.0
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.4.0
setuptools==65.5.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.1
toml==0.10.2
tomlkit==0.12.3
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vdf==3.4
whitenoise==6.6.0
wrapt==1.12.1
zope.event==5.0
zope.interface==6.2
